Move Filter to Nav
