+++ src/components/container/BooksList.js
@@ -2,20 +2,14 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import Book from '../presentation/Book';
-import { removeBook, changeFilter } from '../../actions';
-import CategoryFilter from '../presentation/CategoryFilter';
+import { removeBook } from '../../actions';
 
 const BooksList = (props) => {
   const {
-    books, removeBook, changeFilter, filter,
+    books, removeBook, filter,
   } = props;
   const handleRemoveBook = (book) => {
     removeBook(book.ID);
-  };
-
-  const handleFilterChange = (e) => {
-    const filter = e.target.value;
-    changeFilter(filter);
   };
 
   const filterBooks = (books, filter) => (filter === 'All' ? books
@@ -23,7 +17,6 @@ const BooksList = (props) => {
   const filteredBooks = filterBooks(books, filter);
   return (
     <div className="container">
-      <CategoryFilter handleFilterChange={handleFilterChange} />
       {
         filteredBooks.map((book) => (
           <Book
@@ -43,16 +36,14 @@ const mapStateToProps = (state) => ({
 });
 
 const mapDispatchToProps = (dispatch) => ({
-  removeBook(id) { dispatch(removeBook(id)); },
-  changeFilter(filter) {
-    dispatch(changeFilter(filter));
+  removeBook(id) {
+    dispatch(removeBook(id));
   },
 });
 
 BooksList.propTypes = {
   books: PropTypes.arrayOf(PropTypes.object),
   removeBook: PropTypes.func.isRequired,
-  changeFilter: PropTypes.func.isRequired,
   filter: PropTypes.string.isRequired,
 };
